requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-Blinka==8.39.1
adafruit-circuitpython-ads1x15==2.2.25
adafruit-circuitpython-busdevice==5.2.9
adafruit-circuitpython-connectionmanager==3.0.0
adafruit-circuitpython-requests==3.2.9
adafruit-circuitpython-typing==1.10.3
adafruit-extended-bus==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.63.0
Adafruit-PureIO==1.1.11
certifi==2024.2.2
colorzero==2.0
gpiozero==2.0.1
influxdb_client==1.43.0
pyftdi==0.55.4
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyusb==1.2.1
reactivex==4.0.4
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.16.0
spidev==3.6
sysv-ipc==1.1.0
typing_extensions==4.11.0
urllib3==2.2.1
